requirements.txt
playwright
openai
click
pyyaml
rich
